record query in a celery task
